fix multi deleting products
git-svn-id: https://svn.modified-shop.org/svn/modified-shop/trunk@15703 79ac0487-9261-4f26-a34d-858dee0dcce0

+++ admin/categories.php
@@ -186,7 +186,9 @@ if (xtc_not_null($action)) {
             )
         {
           foreach ($_POST['multi_products'] AS $product_id) {
-            $catfunc->delete_product($product_id, $_POST['multi_products_categories'][$product_id]);
+            if (isset($_POST['multi_products_categories'][$product_id])) {
+              $catfunc->delete_product($product_id, $_POST['multi_products_categories'][$product_id]);
+            }
           }
         }
       }
